+++ src/main/java/com/snepp/backend/SneppController.java
@@ -1,4 +1,4 @@
-package com.snepp.backend.v1.controller;
+package com.snepp.backend;
 
 import com.snepp.backend.v1.model.request.SneppRequest;
 import com.snepp.backend.v1.model.response.SingleSneppResponse;
